fix(gateway): extract program from side-effectful main
Moves makeGatewayProgram, GatewayProgramDeps, and the structured logger into a new side-effect-free program.ts module. main.ts becomes a thin runner; program.test.ts now exercises the program directly without mocking Effect.runPromise. No behavior change.
